ap_news_scraper/requirements.txt
rpaframework==10.5.0
pandas==1.4.0
